Fixed issue with CDC serial_usb not connecting on Windows.
Was a problem with how the ping-pong of the odd/even buffers were handled vs the DATA0/1 field of the BDT for RX packets. Now a setup packet always initializes its current buffer to DATA0 and the opposite buffer to DATA1. On Endpoint setup the even buffer is initialized to DATA0 and the odd to DATA1. This seems correct and is working on endpoint 1 (data comm) for serial_usb.